token CORE

+++ src/main/java/com/vsn/services/interfaces/NodeService.java
@@ -6,6 +6,8 @@ import com.vsn.exceptions.WrongBalanceException;
 import org.web3j.crypto.CipherException;
 
 import java.io.IOException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -15,7 +17,7 @@ public interface NodeService {
 
     Wallet createWallet(User user) throws IOException, CipherException, InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException;
 
-    Double getBalanceWallet(Wallet  wallet) throws IOException;
+    BigDecimal getBalanceWallet(Wallet  wallet) throws IOException;
 
     String sendToAddress(Wallet wallet,String address,double amount) throws WrongBalanceException;
